Remove SSO test user autocreation.
This causes a catch 22 situation when trying to
run initial migrations or schema loads as rake
loads the rails environment and tries to access
the `users` table.  If the CreateUser migration
hasn't run yet, this causes an exception which
prevents the table being created.

We will revisit this in due course.