Add policies to Rummager
This commit changes the publishable concern to send policies to
Rummager. It does this by using the IndexablePresenter added in the
previous commit.

It also exposes the `publish` method to be public as it makes more sense
that we use this to publish things instead of calling save each time.