Don't conflate setup side-effects & expectations
The only reason multiple calls were being made to Publishing API was because
creating the initial policy in the setup for the editing scenario results in a
publish call. Instead of including that in the results expected from the user
action, we can reset the remote calls logged by WebMock during the setup and
verify just what we expect from the actions that will be performed.