Don't verify against schema in features
This is not the right place to verify that the generated JSON for a policy area
or programme is valid according to the schema. Besides which, it is validating
them against the wrong schema - these documents will be pushed to the content
store as a "policy" format, not as a "finder". Unit tests will be added to
validate the output of policy-publisher once we have schema files and examples
for them in govuk-content-schemas.