Testing applicable nations in features
This commit adds tests to both Policy Area and Programmes for testing
the Applicable Nations workflow. I had to modify the pattern matching of
some of the existing steps in order to make them more picky as they were
incorrectly picking up the rest of the steps that I added as they start
off the same.