Extract a Publisher for policies
This encapsulates the things that we expect to happen when we want to put a
policy live on GOV.UK, including publishing parent policies of programmes.

For now, just wire this directly into the Policy model itself so that the test
coverage on Policy still holds. The behaviour has changed slightly in the new
Publisher - parent policies are no longer re-indexed in search. This was more a
result of the previous implementation detail than an actual requirement to
re-index the parent policies.