Untie publishing from Policy model
The act of pushing a Policy live to GOV.UK is now decoupled from the model
itself.

For now, we simply perform the publishing action in the controller after a
successful save. This means we can simplify a bunch of specs and features as we
no longer have to make sure all the various remote requests that are performed
when a Policy is published every time we want to create a Policy instance.