Feature-flag for env-based control of publishing
The Publisher class now behaves differently, depending on the state of a
feature-flag - either publishing a policy live to GOV.UK, or only registering a
"placeholder" content item.

The flag is controlled using the an environment variable called
"ENABLE_FUTURE_POLICIES". This is the same name of the variable used to control
the feature on whitehall.