Fix content_ids on policy seed data
Because of an issue with the database sync between production and preview
(caused by a subtle bug in whitehall and MySQL replication), the content_ids in
the whitehall database dump did not match those in production. This meant that
the seed data generated from the database dump included incorrect content_ids.
This commit updates the seed data to use the correct content_ids. A subsequent
migration will re-seed all the data.