Use FeatureFlag model rather than using ENV vars
We want to be able to change feature flags without having the need to
update the puppet or deployment repos. The current approach of using ENV
variables would require this.

Created a feature flag module which can take a name and a boolean value
and a rake task to update those feature flags.

If you want to update one you can call:

    bundle exec rake feature_flag:set[new-key,true]