Simplify Policies Finder
As items in the Policy finder aren't sorted by the date any content was
tagged to them it makes sense to remove dates on the Policies finder.
Both as result metadata and as a filter. This commit also applies a
default_order to the Policy finder to sort by title.

Ticket:
https://trello.com/c/gZln2whN/274-simplify-policy-finder-of-finders