Add a policy firehose publisher
For government users, we can provide a page that allows them to see all
content tagged to all Policies. To do this we need to reject all
documents which have an empty policy field. We can do this with the
special value of `_MISSING`.

This commit also adds the task to publish this content item to the
Rakefile and a spec to make sure it is valid against the Finder content
schema.