Add detailed format facet
This commit adds the detailed format facet with the list of allowed
values to the Content Item. Initially we wanted to only store these
allowed values in Rummager but that would require Rummager to expand
facets for things other than Objects. It was then decided that storing
these options in the Content Item is a decent solution until this
happens from Rummager.