Remove all migrations which rely on data
Building this App from scratch requires running the migrations. A number
of these migrations ran code which either relied on columns added later
in the migrations, code that has been removed or data that needed to be
present but wasn't added because of an earlier failing migration. This
commit removes any of these migrations and reruns `db:create db:migrate
db:seed` to make sure that the new list of migrations can be ran.