Update Statistics format labels
In order to reduce confusion for both end-users and departmental users,
the Statistics format is being renamed to Official Statistics, and
"Statistics - national statistics" is being renamed to National
Statistics.

[The story](https://trello.com/c/TxBvx7tD/27-make-it-clearer-that-statistics-are-official-or-national-medium)

This commit changes the labels in the policy facets. The "detailed_format"
values will stay the same because those keys are much harder to change
(than the label) due to other things relying on the specific values
(eg the email alert feeds).