Add parent link to email alert signup presenter
Why?:

We want to change email alerting behaviour to use content IDs rather
than slugs. This change will store a reference to the parent policy on
the email alert item stored in the content store. This, in turn, makes
it easier for the email-alert-frontend to handle subscriptions via
email-alert-api by sending the appropriate policy content ID.

How?:

- Update EmailAlertSignupContentItemPresenter#exportable_attributes