Update email alert tags naming convention
email_alert_signup content items carry a reference to the piece of
content they handle subscriptions for. Currently, this reference is
namespaced "details > tags". This is problematic when handling policy
alerts. Any published changes to the email signup page results in the
email-alert-service matching on these tags and sending an alert out to
subscribers (via the email-alert-api). This is not the intended behaviour.

Renaming this field to 'signup_tags' makes its purpose explicit and avoids
this unintentional behaviour.