Remove `decent_exposure`
`decent_exposure` is a gem that automatically "exposes" the a variable
in the views and controller for the model of a controller. While this
saves some typing and boilerplate, it makes the controller less obvious
with "magic" variables coming from nowhere.

The immediate reason to remove it here, aside from making the code
clearer, is that we want to use form objects in the future. This will
decouple the form input from the active record object to allow us to
post (and fetch) from the publishing-api rather than own database.