Remove links from the policy schema
Organisations, people and working groups are no longer stored in our local db.

When viewing or editing policies we now fetch the links from the
publishing api. This way we don't overwrite links that are updated
elsewhere, since policy-publisher is not the authoritive source
of this data.

Temporarily disable features relating to ordering of tags.
publishing-api doesn't preserve the ordering but we think there may be a
need for this; for example, designating a "primary" organisation or
person for a policy.

Add a feature to ensure that the data fetched from the publisher-api
is actually used in the subsequent publish.