Cache people/organisations/working groups
On the policy model, cache the fetched links. This changes behaviour slightly
(see changed spec) but is only actually used for the search indexer, and this
was slowing down saves considerably.

Change ContentItemFetcher to be instantiable so we can reuse fetched data
without repeating queries.

When saving a policy, these collections are used for mapping content_ids to
organisation/people/working group hashes. The way this is computed is not
very efficient because we query all the possible links rather than let the
publishing api do the filtering, so this could be improved further.