Ensure tags lists are sorted
When rendering the widget for entering organisations/people/working
groups, there is a list of values that is filtered as you type.
Ensure all of these are sorted alphabetically so it is possible
to scroll through.

The data we get back from the publishing api currently doesn't
have an ordering to it. We will probably make this an option in future
but for now we can just sort it ourselves before rendering.