Separate organisations into lead and supportng
The content schema for links now has a separate field for lead
organisations, so we can emphasise the primary organisations that the policy
belongs to.

On the model we expose this as lead/supporting so that it's more
obvious whats going on.

Refactor the controller and form to explicitly say what is a PolicyForm
and what is Policy, as this was getting confusing.

Set the link attributes on the form in the from_existing method
instead of having the accessor delegate, so that there is not confusion
between the instance variables of the two objects.

Removed cucumber tests related to ordering. This is no longer applicable,
and we have specs for the new lead organisation field that replaced it.