Add rake task to publish orgs to publishing-api
We are migrating this app to use publishing-api's V2 endpoints. This rake task
is for registering the first organisation of each policy as lead organisation in
the publishing api.

We're also taking the opportunity to republish the whole lot of policy
organisations to publishing api just to make sure its database is in sync with
the one used by policy-publisher.

Part of:
https://trello.com/c/BQDQ824W/468-use-publishing-api-v2-for-policy-publisher