Fetches policy organisations from database
When we merged https://github.com/alphagov/policy-publisher/pull/123  we stopped fetching the `organisation_content_ids` from the database,
so when we ran the rake task on https://github.com/alphagov/policy-publisher/pull/124 we sent a list of empty organisations to publishing-api.

This commit changes the behaviour and fetches the `organisation_content_ids` directly from the database.