Make republish rake task more specific
This rake task is doing four things at the moment: republish the
policy, the email signup page for the policy, the parent policy and
sends the data to rummager. We only want it to send the policy to the
publishing-api. This also allows us to specify the `update_type`, which
will prevent emails from being sent out.