Move tagging lookup to SearchIndexer
To send the proper payload to Rummager, we need to look up the base
path for the items linked to the policy. Currently this happens in the
model. By moving it out to the `SearchIndexer` we can make the model
cleaner and make the code more clear.

The removed unit tests are covered by the tests for the search indexer.