Make user organisation nullable
When we make changes to a regular user in signon, they get synced to the apps.
The user that is used to make the gds-api-adapters request is "SSO Permissions Push User",
and they don't belong to an organisation.

As part of the Push User authentication, something tries to write to the user record,
and it uses null for the organisation. In other apps, this field is defined as nullable
so it works fine, but policy publisher defined it as not nullable and expects the value
to be empty string.

This changes policy publisher to use the same type as other apps, eg content tagger:
https://github.com/alphagov/content-tagger/blob/master/db/schema.rb