Revert the changes made to the Policies finder titles
Revert the changes to the titles made in PR:
https://github.com/alphagov/policy-publisher/pull/167

The titles of the Policies and the All policy content finders were
changed to match each other as part of an A/B test, so that users would
think that they are looking at different versions of the same page.

The A/B test finishes on September 13th, so we can revert the titles.

We are leaving the descriptions in place as they are an improvement.