Change the handling of sub policies
Previously, sub policies could be created, but when you edited them,
they would revert to being a policy, as the parent_policy_ids are not
persisted. Also, the form would show in the policy editing mode, and
be missing the "Part of" field.

As far as I can tell, this change was an unintentional side effect of
e310a1a6dfc73b821f1c5ae672cecbc5894aff6e which splits out link
attributes, but keeps parent_policy_ids as an attribute. The attribute
values are copied to the form via the as_json method, but this will
ignore the associated policies, therefore the default value of [] is
used.

I'm unsure if this is useful functionality to have, but it's kind of
supported, so lets fix it.