Remove the "policy firehose"
This removes the policy firehose (/government/policies/all).

The [policies firehose
finder](https://www.gov.uk/government/policies/allhttps://www.gov.uk/gov
ernment/policies/all) was created by the Elections team during the 2015
General Election, but they didn't have time to test it. So the finder
was deployed to production, but not linked to from any other page on
GOV.UK, with the eventual aim of replacing the original [policies
finder](https://www.gov.uk/government/policies).

As part the of the Government Scrutineers discovery, the policies
firehose finder was tested against the original policies finder.

The results of the test are documented here:
https://gov-uk.atlassian.net/wiki/spaces/GOVUK/pages/168755249/Analysis+
for+Scrutineers

TL;DR The original finder performed slightly better than the newer
firehose finder, so the decision has been taken to remove that finder.

https://trello.com/c/YDda1PqI