Stop using the policy_areas link type
This is causing confusion, as Policy Publisher can use this to
represent the relationship between policies, and in this context, it
doesn't have anything to do with the content in the Publishing API
with a document type of `policy_area`.

The `policy_areas` link type is used elsewhere to represent the
relationship between content (many document types) to policy
areas (either with the document type `policy_area`, or
`placeholder_policy_area`). The potential to do something different
here could cause confusion if it was used.

I don't anticipate removing this to cause any immediate
problems. Either due to issues with Policy Publisher [1] or otherwise,
there are currently no sub policies, so there are no policies using
the `policy_areas` link type in this way.

Given the current work around taxonomies and navigation, I'm not sure
of the future of Policy Publisher, and policies in general, so for
now, I think the thing to do is remove the use of this link type. If
in the future sub policies are used again, and it becomes useful to
represent that relationship in the Publishing API, I don't see any
issues with adding this functionality back, providing a different and
appropriate link type is used.

1: Details in 9695f347b0f2356c13bf504ee38b15b25e031353